remove monolith api module
It was unnecessary. Routes can be defined by any controller, so no
reason to pump everything into an api module/controller/service -- they
are all a part of the api anyway. As the api grows that file would get
inordinately large. Better to fix now rather than when the app gets very
large.

I created a standalone hello-world controller to leave that basic check
that this project's front and backend can communicate.