add '/social-sign-in' to api-freshness strategy
`Refused to execute script from '.chunk.js' because its MIME type
('text/html') is not executable, and strict MIME type checking is
enabled. occurring because of some sort of service worker issue. I think
it is wiping out the module and cannot serve the now non-existent file.
this may or not be the right solution.